refactor: move table into separate module
