+++ rust/src/operations/transaction/mod.rs
@@ -9,7 +9,7 @@ use crate::action::{Action, CommitInfo, DeltaOperation};
 use crate::crate_version;
 use crate::errors::{DeltaResult, DeltaTableError};
 use crate::storage::commit_uri_from_version;
-use crate::table_state::DeltaTableState;
+use crate::table::state::DeltaTableState;
 
 mod conflict_checker;
 #[cfg(feature = "datafusion")]
