+++ rust/src/operations/transaction/state.rs
@@ -24,7 +24,7 @@ use crate::delta_datafusion::{
     get_null_of_arrow_type, logical_expr_to_physical_expr, to_correct_scalar_value,
 };
 use crate::errors::{DeltaResult, DeltaTableError};
-use crate::table_state::DeltaTableState;
+use crate::table::state::DeltaTableState;
 
 impl DeltaTableState {
     /// Get the table schema as an [`ArrowSchemaRef`]
