+++ rust/src/operations/transaction/test_utils.rs
@@ -1,7 +1,7 @@
 #![allow(unused)]
 use super::{prepare_commit, try_commit_transaction, CommitInfo};
 use crate::action::{Action, Add, DeltaOperation, MetaData, Protocol, Remove, SaveMode};
-use crate::table_state::DeltaTableState;
+use crate::table::state::DeltaTableState;
 use crate::{
     DeltaTable, DeltaTableBuilder, DeltaTableMetaData, Schema, SchemaDataType, SchemaField,
 };
