+++ rust/src/table/state_arrow.rs
@@ -16,9 +16,9 @@ use arrow_array::{
 use arrow_schema::{DataType, Field, Fields, TimeUnit};
 use itertools::Itertools;
 
+use super::state::DeltaTableState;
 use crate::action::{ColumnCountStat, ColumnValueStat, Stats};
 use crate::errors::DeltaTableError;
-use crate::table_state::DeltaTableState;
 use crate::SchemaDataType;
 use crate::SchemaTypeStruct;
